Fix network stats with Docker 1.9+
Starting with API 1.21, the network statistics consist in a
dictionary with the network interface as key and the actual stats as
values. The commit introduce two methods to compute the network stats
depending on what's returned by the server. If a single `network` is
returned, the behaviour isn't changed. If multiple `networks` are
returned, the sender will compute the sum of all the interfaces'
traffic.